Split useBackendRender function
